chore: refactor the qa service into modules
